requirements.txt
# Core libs
pillow
psycopg2
django>=2.0,<=2.0.7
django-cors-headers
djangorestframework>=3.6.5,<=3.8.2
django-widget-tweaks
requests
markdown<=2.6.11
factory-boy==2.11.1
pytest-django==3.4.2
django-filter==2.0.0
django-haystack==2.8.1
elasticsearch==2.4.1
drf-extensions==0.4.0
drf-haystack==1.8.4
celery==4.2.1
django-rest-auth==0.9.3
django-allauth

# Dev tools
django-extensions
django-rest-swagger
